First Tests & Validator existing UserMail & Reservation productAmount & Creating Command to clear unused baskets

+++ src/Controller/BasketController.php
@@ -2,9 +2,11 @@
 
 namespace App\Controller;
 
+use App\Entity\User;
 use App\Repository\BasketProductRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 
 class BasketController extends AbstractController
@@ -24,8 +26,9 @@ class BasketController extends AbstractController
      */
     public function basket(): Response
     {
+        /** @var User $user $user */
         $user = $this->getUser();
-        $basketProducts = $this->basketProductRepository->findAllProductsForUser($user);
+        $basketProducts = $this->basketProductRepository->findAllBasketProductsForUser($user);
 
         return $this->render(
             'basket/index.html.twig',
@@ -43,6 +46,9 @@ class BasketController extends AbstractController
     public function basketDelete($id): Response
     {
         $basketProduct = $this->basketProductRepository->find($id);
+        if (null === $basketProduct) {
+            throw new NotFoundHttpException();
+        }
         $this->basketProductRepository->delete($basketProduct);
 
         return $this->redirectToRoute('app_basket');
